# Backend/SHOP_RUNNER/SHOP_RUNNER/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SHOP_RUNNER, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Name",
            "required": true,
            "type": "string",
            "maxLength": 255,
            "minLength": 3
          },
          {
            "in": "formData",
            "name": "price",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "qty",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "category_id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "user_id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "gender_id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "brand_id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "size_id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "color_id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/get-by-id": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/update": {
      "put": {
        "tags": [
          "Product"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "Id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Name",
            "required": true,
            "type": "string",
            "maxLength": 255,
            "minLength": 3
          },
          {
            "in": "formData",
            "name": "price",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "Thumbnail",
            "type": "file"
          },
          {
            "in": "formData",
            "name": "qty",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "categoryId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "genderId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "brandId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "sizeId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "colorId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/Search": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/related": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/Paginate": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/Filter": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "to",
            "type": "number",
            "format": "double",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "from",
            "type": "number",
            "format": "double",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "category",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "gender",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "brand",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "size",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "color",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/Sorting": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sortBy",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "page",
            "type": "integer",
            "format": "int32",
            "default": 1,
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/delete": {
      "delete": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Product/toggle-product": {
      "post": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "product_id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Staff/create-acc-staff": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Staff_register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Staff/change-password": {
      "post": {
        "tags": [
          "Staff"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Staff_changePass"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Staff/toggle-user": {
      "post": {
        "tags": [
          "Staff"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Staff_changePass": {
      "required": [
        "id",
        "password_new",
        "password_old"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "password_old": {
          "minLength": 6,
          "type": "string"
        },
        "password_new": {
          "minLength": 6,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Staff_register": {
      "required": [
        "confirmPassword",
        "fullName",
        "password"
      ],
      "type": "object",
      "properties": {
        "fullName": {
          "minLength": 6,
          "type": "string"
        },
        "password": {
          "minLength": 6,
          "type": "string"
        },
        "confirmPassword": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme (\"bearer {token}\")"
    }
  }
}